Deconstruction of d3drender for world objects
